fixup! Reuse code
